# api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: DNS Frontdoor Backend
  description: "This server is for the DNS Frontdoor"
  contact:
    email: pranav.bodawala@telus.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://localhost:3000
...
paths:
...
  /{slug}:
    get:
      tags:
        - shortener
      summary: "Redirect to the long url"
      operationId: Redirect
      parameters:
        - name: slug
          in: path
          description: slug (aka id) of the long url
          example: a1b2c3
          required: true
          schema:
            type: string
      responses:
        200:
          description: Redirected to the long url
          content:
            application/json:
              example: "Redirect to long url"
        404:
          description: Cannot find the slug in the database
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "a1b2c3 does not exist"
                  code:
                    type: number
                    enum: [404]
...